Add test & fix bug
